feat(redis): seperate the creation of cache
