+++ pkg/cache/module.go
@@ -6,8 +6,17 @@ import (
 )
 
 var (
+	// RedisRingModule for fx.
+	RedisRingModule = fx.Provide(redis.NewRing)
+
+	// RedisOptionsModule for fx.
+	RedisOptionsModule = fx.Provide(redis.NewOptions)
+
+	// RedisCacheModule for fx.
+	RedisCacheModule = fx.Provide(redis.NewCache)
+
 	// RedisModule for fx.
-	RedisModule = fx.Provide(redis.NewCache)
+	RedisModule = fx.Options(RedisRingModule, RedisOptionsModule, RedisCacheModule)
 
 	// Module for fx.
 	Module = fx.Options(RedisModule)
